+++ lib/index.ts
@@ -1,3 +1,3 @@
 export { LibraClient, LibraNetwork } from './client';
 export { LibraWallet } from './wallet';
-export { Account, AccountState } from './wallet/accounts';
+export { Account, AccountState } from './wallet/Accounts';
